fix(auditoria capa 6): metricas derivadas — price-check porciones, TRIM categoria, Sale parseFloat [A1-C1]
Capa 6 — Riesgo Medio. Tres fixes en métricas derivadas:

1. A1-C1 — intelligence/price-check (intelligence.routes.js):
   - SELECT ahora recupera r.porciones (antes ausente).
   - Se acumula costeLote (Σ ingrediente × cantidad) y se divide por porciones
     para obtener coste por porción antes de calcular food cost%.
   - Antes: receta de N porciones reportaba food cost ×N (multiplicado, no
     dividido). Recetas con porciones=4 a 12€ y coste lote 6€ reportaban
     food cost 50% (incorrecto, ALERTA falsa); ahora reportan 12.5% (correcto).
   - Alineado con frontend (calcularCosteRecetaCompleto) y getRecipeCostBase.

2. TRIM en SQL bucketing categoría (analysis.routes.js, analytics.routes.js,
   chatService.js):
   - LOWER(COALESCE(r.categoria, '')) → LOWER(TRIM(COALESCE(r.categoria, '')))
   - Sincroniza el SQL con el helper JS classifyCategoria que ya hace .trim().
   - Cubre el [HALLAZGO_NUEVO] documentado en MICRO_INFORME_CAPA_5 sobre
     categorías con espacios al final que no matcheaban.

3. Sale.js (domain entity) — parseInt → parseFloat:
   - this.quantity = parseFloat(data.cantidad || data.quantity) || 0.
   - Mismo fix que Capa 4 monthly.routes.js (A1-C2): preserva medias porciones.
   - validate() sigue rechazando quantity <= 0; medias porciones positivas
     pasan validación correctamente.

Verificación SQL columnas (lección hotfix Capa 3):
  - recetas.porciones (init.js:85) ✓
  - recetas.categoria (init.js:83) ✓

Tests: lint 0 errores, unit+services 35/35, intelligence-endpoints 5/5,
menu-engineering-filter 3/3, recipe-crud-food-cost 5/5, monthly-balance-pnl
3/3, daily-summary-pnl 3/3.

Pendiente Capa 7 (Bajo riesgo) frontend:
- Locale dinámico es-ES hardcodeado.
- Re-parse cm() formateado en balance/index.js.
- helpers.js parseInt en proyección consumo.
- S7 thresholds frontend.

+++ src/domain/entities/Sale.js
@@ -13,7 +13,11 @@ class Sale {
         // Soportar todos los formatos: receta_id, recipeId, recetaId
         this.recipeId = data.receta_id || data.recipeId || data.recetaId;
         this.recipeName = data.receta_nombre || data.recipeName || null;
-        this.quantity = parseInt(data.cantidad || data.quantity) || 0;
+        // 🔒 Auditoría A1-C2 (Capa 6): parseFloat preserva medias porciones (0.5).
+        //    Antes parseInt truncaba ½ ración a 0 y la validación quantity > 0 la
+        //    rechazaba como venta. Alineado con monthly.routes.js (Capa 4) que ya
+        //    usa parseFloat al leer SUM(v.cantidad) desde la BD.
+        this.quantity = parseFloat(data.cantidad || data.quantity) || 0;
         this.unitPrice = parseFloat(data.precio_unitario || data.unitPrice) || 0;
         this.total = parseFloat(data.total) || 0;
         this.date = data.fecha || data.date || new Date();
